feat: lock appointment deletion to testing users

+++ server/actions/deleteAppointment.bypass.ts
@@ -1,10 +1,19 @@
 "use server";
 
 import { AppointmentsRepository } from "@/server/adapters/appwrite/appointmentsRepository";
+import { isTestingUser } from "@/server/adapters/appwrite/isTestingUser";
+import { ForbiddenInTestingFailure } from "@/server/useCases/shared/failures/forbiddenInTestingFailure";
 
 export const deleteAppointment = async (formData: FormData) => {
-  const repository = new AppointmentsRepository();
-  const result = await repository.delete(formData.get("id") as string);
+  let result;
+
+  if (await isTestingUser()) {
+    result = new ForbiddenInTestingFailure();
+  } else {
+    const repository = new AppointmentsRepository();
+    result = await repository.delete(formData.get("id") as string);
+  }
+
   const plainObject = Object.assign({}, result);
   return plainObject;
 };
